There can only be one. RIFContextManager, that is. And now there is.
