+++ rifServices/src/main/resources/RIFLogger.properties
@@ -90,7 +90,7 @@ rifServices.restfulWebServices.pg.PGSQLRIFStudySubmissionWebServiceResource=true
 rifServices.dataStorageLayer.common.ProductionStudyRetrievalService=false
 
 rifServices.dataStorageLayer.common.ResultsQueryManager=true
-rifServices.dataStorageLayer.pg.PGSQLRIFContextManager=true
+rifServices.dataStorageLayer.common.RIFContextManager=true
 rifServices.dataStorageLayer.common.BaseSQLManager=true
 rifServices.dataStorageLayer.pg.PGSQLAgeGenderYearManager=true
 rifServices.dataStorageLayer.pg.PGSQLAgeGenderYearManager=true
